+++ src/pages/restaurants/Regions/RegionsContainer.jsx
@@ -1,9 +1,7 @@
 import { useDispatch, useSelector } from 'react-redux';
 
-import {
-  selectRegion,
-  loadRestaurants,
-} from '../../../store/actions';
+import { selectRegion } from '../../../store/actions';
+import { loadRestaurants } from '../../../store/async-actions';
 
 import { get } from '../../../utils';
