+++ src/store/async-actions.js
@@ -1,0 +1,40 @@
+import {
+  fetchRegions,
+  fetchCategories,
+  fetchRestaurants,
+} from '../services/api';
+
+import {
+  setCategories,
+  setRegions,
+  setRestaurants,
+} from './actions';
+
+export function loadInitialData() {
+  return async (dispatch) => {
+    const regions = await fetchRegions();
+    dispatch(setRegions(regions));
+
+    const categories = await fetchCategories();
+    dispatch(setCategories(categories));
+  };
+}
+
+export function loadRestaurants() {
+  return async (dispatch, getState) => {
+    const {
+      selectedRegion: region,
+      selectedCategory: category,
+    } = getState();
+
+    if (!region || !category) {
+      return;
+    }
+
+    const restaurants = await fetchRestaurants({
+      regionName: region.name,
+      categoryId: category.id,
+    });
+    dispatch(setRestaurants(restaurants));
+  };
+}
